app/requirements.txt
noisereduce
boto3
ffmpeg-python